+++ src/Presentation/SpaWeb/src/App.js
@@ -6,8 +6,8 @@ import { Container } from "react-bootstrap";
 import { ToastContainer } from "react-toastify";
 import { NetworkError } from "./components/Error/NetworkError";
 import { NotFound } from "./components/Error/NotFound";
-import { Register } from "./components/Register/Register";
-import { Login } from "./components/Login/Login";
+import { Register } from "./components/Account/Register/Register";
+import { Login } from "./components/Account/Login/Login";
 import { ProvideAuth } from "./utils/hooks/authHook";
 import { List } from "./components/Items/List/List";
 import { Details } from "./components/Items/Details/Details";
